Move files out of helpers directory which is meant for handlebar helper functions
