+++ app/components/input/input-date.js
@@ -1,6 +1,6 @@
 import Component from '@ember/component';
 import { dateGreaterNow } from 'kursausschreibung/framework/date-helpers';
-import { formDanger } from 'kursausschreibung/helpers/form';
+import { formDanger } from 'kursausschreibung/framework/form-helpers';
 
 export default Component.extend({
     change() {
